Fixed issue #50.
After splitting a cookie on '=', we assume there can be on or more parts,
rather than assuming a pair. We check the cookie name by shif()ing off the first
part, and then join() the remaining parts to get the value. This also handles
the IE case where an empty cookie will have no '=' at all, and we end up with
and empty string as before.