Test for issue #88, loads a second copy of the plugin into a fake environment
IE has a bug where it sometimes generates empty values in document.cookie.
This is a test which illustrates how we choke on them.

The fake environment is set up using eval() inside a closure,
fetching the script using synchronous XHR.  Setup errors are
displayed using an alert().  It might want refactoring in future,
but this is my best draft so far.  (Imagine what the earlier
ones looked like ;).